Se optimiza el archivo server.js y se crea el archivo weebhook.js que manejará las peticiones que se realicen por parte del bot
